requirements.txt
# Core
numpy
scipy
pandas

# ML
scikit-learn

# Image processing
Pillow

# Web API
fastapi
uvicorn[standard]

# Monitoring / logging (opsional, tapi berguna)
python-multipart

# Model persistence
joblib
